Warn instead of fail when keychain is unreachable (#78)
Warn the user the keychain is unreachable instead of raising an error.

- When the master password is set for the first time (the DB was absent)
or configured to be remembered (`dcli configure save-master-password
true`), this will disable the master password (encrypted using the local
key) save in the DB (that is then useless because the local key can only
be retrieved from the master password). A warning is displayed, so the
user can retry using the keychain.
- When the keychain suddenly starts to fail (`save-master-password` set
to `true` but `getLocalKey()` returns an error) or the master password
encrypted using the local key is absent from the DB, then the keychain
is not disabled, but a warning is displayed, so the user knows what is
happening and how to disable the keychain.

Because the keychain unavailability does not trigger errors anymore, it
has been possible to postpone its access. Thus, this PR resolves #55.

If you think of better warnings, don't hesitate to edit them!