Improve filters and add JSON output for the `note` command
Previous filters were only allowing to fetch records by title.
New approach to filters allow to specify which field to look at, and
mix and match multiple filters. All is applied with the `or` logic,
and keeps the previous logic working so no breaking change.

The `--output json` parameter allows to return the notes in JSON.

e.g.
  `dcli n id=xxxxxx`
    will return any entry for which the `id` matches `xxxxxx`

  `dcli n content="some content" title=mytitle`
    will return any entry for which the `content` matches
    `some content`, or the `title` matches `mytitle`

  `dcli n content,title="my data"`
    will return any entry for which either the `content` or
    the `title` matches `my data`

  `dcli n --output json`
    returns all the notes in a JSON list

  `dcli n --output json title=mytitle`
    returns all the matching notes in a JSON list

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>